# Optimize encoding by using CBOR lib's StreamEncoder