Closes #794 (epic)
Closes #796
Closes #807

## Description

- Use streaming mode for encoding to eliminate intermediate objects being created when encoding large composites and dictionaries.

- Use low-level API for encoding to reduce CBOR library's internal calls to Go's reflect package (to boost speed).

- Modify go.mod to use fxamacker/cbor branch feature/stream-mode before it's merged to master.

## Cumulative Benchmark Comparisons

These comparisons are cumulative vs commit 3bbc8c8, so they include performance gains from switching to v4 storage format (issue #738) and other optimizations by the Flow team.

<details><summary>Original PR before hard coding tag numbers and array lengths</summary>

```
name                                               old time/op    new time/op    delta
EncodeCBOR/comp_v2_96d7e06eaf4b3fcf_14850bytes-4      885µs ± 1%     151µs ± 1%  -82.96%  (p=0.000 n=10+10)
EncodeCBOR/comp_v3_99dc360eee32dcec_160bytes-4       20.7µs ± 1%     4.4µs ± 1%  -78.93%  (p=0.000 n=9+10)
EncodeCBOR/comp_v3_b52a33b7e56868f6_139bytes-4       15.9µs ± 1%     3.3µs ± 0%  -79.32%  (p=0.000 n=10+9)
EncodeCBOR/comp_v3_d99d7e6b4dad41e1_776155bytes-4    35.0ms ± 1%     6.9ms ± 1%  -80.22%  (p=0.000 n=9+10)

name                                               old alloc/op   new alloc/op   delta
EncodeCBOR/comp_v2_96d7e06eaf4b3fcf_14850bytes-4      247kB ± 1%      75kB ± 0%  -69.49%  (p=0.000 n=10+10)
EncodeCBOR/comp_v3_99dc360eee32dcec_160bytes-4       4.85kB ± 2%    1.26kB ± 0%  -73.92%  (p=0.000 n=10+10)
EncodeCBOR/comp_v3_b52a33b7e56868f6_139bytes-4       3.78kB ± 1%    0.86kB ± 0%  -77.15%  (p=0.000 n=10+10)
EncodeCBOR/comp_v3_d99d7e6b4dad41e1_776155bytes-4    7.47MB ± 4%    1.96MB ± 0%  -73.72%  (p=0.000 n=9+10)

name                                               old allocs/op  new allocs/op  delta
EncodeCBOR/comp_v2_96d7e06eaf4b3fcf_14850bytes-4      5.02k ± 0%     0.62k ± 0%  -87.66%  (p=0.000 n=10+10)
EncodeCBOR/comp_v3_99dc360eee32dcec_160bytes-4          103 ± 1%        19 ± 0%  -81.50%  (p=0.000 n=10+10)
EncodeCBOR/comp_v3_b52a33b7e56868f6_139bytes-4         76.0 ± 0%      14.0 ± 0%  -81.58%  (p=0.000 n=10+10)
EncodeCBOR/comp_v3_d99d7e6b4dad41e1_776155bytes-4      141k ± 0%        6k ± 0%  -95.79%  (p=0.000 n=8+10)

--------------------------------------------

name                                            old time/op    new time/op    delta
EncodeCBOR/link_v3_2392f05c3b72f235_167bytes-4    20.3µs ± 1%     3.4µs ± 0%  -83.41%  (p=0.000 n=10+10)
EncodeCBOR/link_v3_3a791fe1b8243e73_192bytes-4    20.6µs ± 1%     3.4µs ± 1%  -83.66%  (p=0.000 n=10+10)

name                                            old alloc/op   new alloc/op   delta
EncodeCBOR/link_v3_2392f05c3b72f235_167bytes-4    5.18kB ± 0%    1.17kB ± 0%  -77.44%  (p=0.000 n=10+10)
EncodeCBOR/link_v3_3a791fe1b8243e73_192bytes-4    5.50kB ± 0%    1.15kB ± 0%  -79.06%  (p=0.000 n=10+10)

name                                            old allocs/op  new allocs/op  delta
EncodeCBOR/link_v3_2392f05c3b72f235_167bytes-4       101 ± 0%        13 ± 0%  -87.13%  (p=0.000 n=10+10)
EncodeCBOR/link_v3_3a791fe1b8243e73_192bytes-4       101 ± 0%        13 ± 0%  -87.13%  (p=0.000 n=10+10)
```

</details>

After hard coding tag number and array length where feasible (thanks @turbolent).
```
name                                               old time/op    new time/op    delta
EncodeCBOR/comp_v2_96d7e06eaf4b3fcf_14850bytes-4      898µs ± 1%     147µs ± 1%  -83.62%  (p=0.000 n=10+9)
EncodeCBOR/comp_v3_99dc360eee32dcec_160bytes-4       21.0µs ± 1%     4.3µs ± 1%  -79.58%  (p=0.000 n=10+9)
EncodeCBOR/comp_v3_b52a33b7e56868f6_139bytes-4       16.1µs ± 0%     3.2µs ± 1%  -80.23%  (p=0.000 n=10+10)
EncodeCBOR/comp_v3_d99d7e6b4dad41e1_776155bytes-4    35.5ms ± 1%     7.0ms ± 0%  -80.44%  (p=0.000 n=8+10)

name                                               old alloc/op   new alloc/op   delta
EncodeCBOR/comp_v2_96d7e06eaf4b3fcf_14850bytes-4      247kB ± 2%      75kB ± 0%  -69.45%  (p=0.000 n=10+10)
EncodeCBOR/comp_v3_99dc360eee32dcec_160bytes-4       4.84kB ± 1%    1.26kB ± 0%  -73.90%  (p=0.000 n=10+10)
EncodeCBOR/comp_v3_b52a33b7e56868f6_139bytes-4       3.78kB ± 1%    0.86kB ± 0%  -77.17%  (p=0.000 n=10+10)
EncodeCBOR/comp_v3_d99d7e6b4dad41e1_776155bytes-4    7.48MB ± 4%    1.96MB ± 0%  -73.76%  (p=0.000 n=9+10)

name                                               old allocs/op  new allocs/op  delta
EncodeCBOR/comp_v2_96d7e06eaf4b3fcf_14850bytes-4      5.02k ± 1%     0.62k ± 0%  -87.65%  (p=0.000 n=10+10)
EncodeCBOR/comp_v3_99dc360eee32dcec_160bytes-4          103 ± 0%        19 ± 0%  -81.55%  (p=0.002 n=8+10)
EncodeCBOR/comp_v3_b52a33b7e56868f6_139bytes-4         76.0 ± 0%      14.0 ± 0%  -81.58%  (p=0.000 n=10+10)
EncodeCBOR/comp_v3_d99d7e6b4dad41e1_776155bytes-4      141k ± 0%        6k ± 0%  -95.79%  (p=0.000 n=9+10)

---------------------

name                                            old time/op    new time/op    delta
EncodeCBOR/link_v3_2392f05c3b72f235_167bytes-4    20.5µs ± 1%     3.3µs ± 1%  -84.00%  (p=0.000 n=10+10)
EncodeCBOR/link_v3_3a791fe1b8243e73_192bytes-4    20.9µs ± 1%     3.3µs ± 1%  -84.32%  (p=0.000 n=10+10)

name                                            old alloc/op   new alloc/op   delta
EncodeCBOR/link_v3_2392f05c3b72f235_167bytes-4    5.18kB ± 0%    1.17kB ± 0%  -77.44%  (p=0.000 n=10+10)
EncodeCBOR/link_v3_3a791fe1b8243e73_192bytes-4    5.50kB ± 0%    1.15kB ± 0%  -79.06%  (p=0.002 n=8+10)

name                                            old allocs/op  new allocs/op  delta
EncodeCBOR/link_v3_2392f05c3b72f235_167bytes-4       101 ± 0%        13 ± 0%  -87.13%  (p=0.000 n=10+10)
EncodeCBOR/link_v3_3a791fe1b8243e73_192bytes-4       101 ± 0%        13 ± 0%  -87.13%  (p=0.000 n=10+10)
```

## Caveats

This requires fxamacker/cbor branch feature-stream-mode, not the master branch.

I initially preferred 2 PRs for this but it seems easier to review because the same files are changed.  I'm open to suggestions if we need to split this up into multiple PR.

______

For contributor use:

- [x] Targeted PR against `master` branch
- [x] Linked to Github issue with discussion and accepted design OR link to spec that describes this work
- [x] Code follows the [standards mentioned here](https://github.com/onflow/cadence/blob/master/CONTRIBUTING.md#styleguides)
- [x] Updated relevant documentation 
- [x] Re-reviewed `Files changed` in the Github PR explorer
- [x] Added appropriate labels